bdcn/requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
fire==0.2.0
idna==2.7
imageio==2.8.0
imageio-ffmpeg==0.4.1
imutils==0.5.4
kiwisolver==1.1.0
lmdb==0.98
matplotlib==3.1.1
moviepy==1.0.0
natsort==7.0.1
nltk==3.4.4
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pipenv==2018.11.26
proglog==0.1.9
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
requests==2.20.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
shape==0.1.1
six==1.14.0
soupsieve==2.0
termcolor==1.1.0
tqdm==4.43.0
urllib3==1.24
virtualenv==16.6.0
virtualenv-clone==0.5.3